Authorization

+++ Services/CharacterService.cs
@@ -1,11 +1,14 @@
 ﻿using AutoMapper;
+using DungeonsAndDragonsCharacter.API.Authorization;
 using DungeonsAndDragonsCharacter.API.Entities;
 using DungeonsAndDragonsCharacter.API.Exceptions;
 using DungeonsAndDragonsCharacter.API.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace DungeonsAndDragonsCharacter.API.Services
@@ -25,13 +28,17 @@ namespace DungeonsAndDragonsCharacter.API.Services
         private readonly CharacterDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<CharacterService> _logger;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IGamerContextService _gamerContextService;
 
 
-        public CharacterService(CharacterDbContext dbContext , IMapper mapper, ILogger<CharacterService> logger  )
+        public CharacterService(CharacterDbContext dbContext , IMapper mapper, ILogger<CharacterService> logger  ,
+            IAuthorizationService authorizationService, IGamerContextService gamerContextService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
 
         }
 
@@ -63,6 +70,7 @@ namespace DungeonsAndDragonsCharacter.API.Services
         public int Create(CreateCharacterDto dto)
         {
             var character = _mapper.Map<Character>(dto);
+            character.CreatedById = _gamerContextService.GetGamerId;
             _dbContext.Characters.Add(character);
             _dbContext.SaveChanges();
 
@@ -80,6 +88,14 @@ namespace DungeonsAndDragonsCharacter.API.Services
             if (character is null)
                 throw new NotFoundException("Character not found");
 
+            var authorizationResult = _authorizationService.AuthorizeAsync(_gamerContextService.Gamer, character,
+              new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
+
             _dbContext.Characters.Remove(character);
             _dbContext.SaveChanges();
 
@@ -94,6 +110,13 @@ namespace DungeonsAndDragonsCharacter.API.Services
             if (character is null)
                 throw new NotFoundException("Character not found");
 
+            var authorizationResult = _authorizationService.AuthorizeAsync(_gamerContextService.Gamer, character,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _dbContext.Characters.Update(character);
             _dbContext.SaveChanges();
